refactor: use Topics component in Card and ProjectDetail
