+++ packages/gatsby-theme-project-portal/src/components/ProjectDetail.tsx
@@ -2,17 +2,16 @@ import React, { FunctionComponent } from "react"
 import moment from "moment"
 
 import {
+  Accordion,
   CollaboratorDetails,
   ContactType,
-  Feature,
+  Topics,
   MainContact,
   ProjectTeam,
   SectionOfItem,
   ShareProject,
   TopicType,
 } from "."
-
-import { Accordion } from "./Accordion"
 
 import { statusOutput, isNA, isEmpty } from "../utils"
 
@@ -134,11 +133,7 @@ export const ProjectDetailLayout: FunctionComponent<
           <section className="flex flex-wrap items-start py-6 m-responsive gap-x-10 gap-y-4">
             {!isEmpty(topics) && (
               <div className="text-tag mt-2">
-                <Feature
-                  label="Topics"
-                  className="bg-topics"
-                  value={topics.map((topic) => topic.title)}
-                />
+                <Topics topics={topics} />
               </div>
             )}
           </section>
